Slice Decedent identifier by profile
Since DecedentID does not fix the `type` property, we can't include it
in a slicing that uses `type` as a discriminator path.  Instead, slice by
profile.

NOTE: This swaps one issue for another in the FHIR IG validation.  While
I *think* this approach is valid, the validator emits an error because
it doesn't support validation of profile-based slices yet.